fix(agent): 修复鼠标滚轮事件无法工作的Bug
根因：Go MouseWheelPayload 的 DeltaX/DeltaY 声明为 int，但浏览器
WheelEvent.deltaX/deltaY 是 double 类型，触控板和 macOS 产生非整数
值时 json.Unmarshal 静默失败，滚轮事件被完全丢弃。

修复内容：
- datachannel.go: DeltaX/DeltaY 从 int 改为 float64
- mouse.go: 重写 MouseWheelCmd，支持滚动幅度(xdotool --repeat)、
  水平滚动(button 6/7)，去掉无效的 click 0
- main.go: 适配新返回类型，添加反序列化错误日志
- VideoCanvas.tsx: 添加 deltaMode 归一化(像素÷40→行,页×30→行)
- signaling.ts: MouseWheelPayload 添加 delta_mode 字段
- 新增单元测试覆盖整数/浮点/水平/对角线滚动场景

+++ evdi-agent/pkg/webrtc/datachannel.go
@@ -27,11 +27,14 @@ type MouseButtonPayload struct {
 }
 
 // MouseWheelPayload represents a mouse scroll event.
+// DeltaX/DeltaY use float64 because browser WheelEvent.deltaX/deltaY are double
+// and frequently non-integer (e.g., trackpad: 3.5, macOS: 4.2).
+// Using int caused silent JSON unmarshal failures for fractional values.
 type MouseWheelPayload struct {
-	DeltaX int `json:"delta_x"`
-	DeltaY int `json:"delta_y"`
-	X      int `json:"x"`
-	Y      int `json:"y"`
+	DeltaX float64 `json:"delta_x"`
+	DeltaY float64 `json:"delta_y"`
+	X      int     `json:"x"`
+	Y      int     `json:"y"`
 }
 
 // KeyPayload represents a keyboard key event.
